refactor: rename iris engine files and design-lab routes
- src/lib/iris-mechanism.ts → iris-kinematics.ts
- src/lib/aperture.ts → iris-pheno-kinematics.ts
- design-lab/aperture-v2/ → design-lab/iris/
- design-lab/logo-mark/ → design-lab/iris-pheno/

Update all imports, path strings, and inline comments to match.
